Remove glu_mangle.h
Long ago when GLU lived in the Mesa repository, there was "name
mangling" support that allowed you to build Mesa and GLU with all
exported symbols' names "mangled". That is, the symbol names would
typically be prefixed with an 'm' so that they would not collide with
the system's OpenGL implementation.

When GLU was split out into its own repository in 2012 we didn't bring
over the name mangling support. No one has noticed this lack of
functionality since, to my knowledge.

Remove the name mangling header (which was unexpectedly licensed under
the GPL).